Verschillende tabMeerdere tabellen
Prompt:
It works and I updated the csv files with the actual content I want.
Please leave them as they are, I also added git for versioning. It
works, the first version is oke. I have a request:n the original xls
file I made my csv file from I had multiple tables which are all
converted to one table. This makes sense because it is not clear which
are headers from the different tables, the headers from the different
tables were:Specifications cable (A)Specifications fiberStandard
compliances